Move download_blob_to_temp to common logic
