+++ f/common_logic/azure_operations.script.lock
@@ -1,0 +1,12 @@
+azure-storage-blob==12.19.0
+azure-core==1.29.5
+typing_extensions==4.8.0
+requests==2.31.0
+urllib3==2.0.7
+certifi==2023.11.17
+charset-normalizer==3.3.2
+idna==3.6
+cryptography==42.0.8
+cffi==1.16.0
+pycparser==2.21
+isodate==0.7.2
