+++ f/common_logic/azure_operations.script.yaml
@@ -1,0 +1,11 @@
+summary: 'Shared logic for Azure operations'
+description: ''
+lock: '!inline f/common_logic/azure_operations.script.lock'
+has_preprocessor: false
+kind: script
+no_main_func: true
+schema:
+  $schema: 'https://json-schema.org/draft/2020-12/schema'
+  type: object
+  properties: {}
+  required: []
